backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
matplotlib==3.9.2
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
